+++ raster_axes/__init__.py
@@ -1,1 +1,3 @@
-from .raster_axes import *
+from .scatter_density_artist import *
+from .scatter_density_axes import *
+
