+++ raster_axes/scatter_density_artist.py
